+++ src/Products/PageTemplates/tests/testExpressions.py
@@ -4,6 +4,7 @@ import unittest
 
 from six import text_type
 
+from AccessControl import safe_builtins
 from zope.component.testing import PlacelessSetup
 
 
@@ -199,6 +200,12 @@ class EngineTestsBase(PlacelessSetup):
                        IUnicodeEncodingConflictResolver)
         self.assertEqual(ec.evaluate(expr), u'äüö')
 
+    def test_builtin_in_path_expr(self):
+        ec = self._makeContext()
+        self.assertIs(ec.evaluate('True'), True)
+        self.assertIs(ec.evaluate('False'), False)
+        self.assertIs(ec.evaluate('nocall: test'), safe_builtins["test"])
+
 
 class UntrustedEngineTests(EngineTestsBase, unittest.TestCase):
 
@@ -208,6 +215,15 @@ class UntrustedEngineTests(EngineTestsBase, unittest.TestCase):
 
     # XXX:  add tests that show security checks being enforced
 
+    def test_open_in_path_expr(self):
+        ec = self._makeContext()
+        with self.assertRaises(KeyError):
+            ec.evaluate("nocall:open")
+
+    def test_list_in_path_expr(self):
+        ec = self._makeContext()
+        self.assertIs(ec.evaluate('nocall: list'), safe_builtins["list"])
+
 
 class TrustedEngineTests(EngineTestsBase, unittest.TestCase):
 
@@ -216,6 +232,14 @@ class TrustedEngineTests(EngineTestsBase, unittest.TestCase):
         return createTrustedZopeEngine()
 
     # XXX:  add tests that show security checks *not* being enforced
+
+    def test_open_in_path_expr(self):
+        ec = self._makeContext()
+        self.assertIs(ec.evaluate("nocall:open"), open)
+
+    def test_list_in_path_expr(self):
+        ec = self._makeContext()
+        self.assertIs(ec.evaluate('nocall: list'), list)
 
 
 class UnicodeEncodingConflictResolverTests(PlacelessSetup, unittest.TestCase):